Add start and reset button and functions for them.
There is now a start button that activates a
function that creates a grid with the number
specified by an input variable.

There is a reset button that for now only reactivates
the start button to play another game.

Reset button lacks complete funcionality.

Basic CSS was added to make website more easy to
understand while working with the JS code logic.