Modify Grid Creation Approach and CSS flexbox.
The previous way to make a grid had a defect
consisting of the cells not being evenly distributed.
That is, for example,
instead of being 16x16 and taking all the space of the
grid, they would be 8x32 and leave blank space.

The new approach uses a new function createGridItems()
that creates a defined number of rows (divs) and a defined
number of cells (divs) within those rows. This way,
the grid is easier to manipulate using Flexbox.

After some tinkering with flexbox, the grid is now
working as expected.