Add change color on mouseover functionality.
In order to be able to "draw" using the grid,
which is the main pupose of this project,
a new event listener was added within the second
for loop in the createGridItems function. This
event listener changes the color of the cell that
the user hovers the mouse over. The color stays
changed so it looks as if the user is drawing.

Note: hover pseudoclass in CSS just changed
the color as long as the mouse was over the cell.
The event listener had to be added within a for
loop so it was applied to each and every cell.